move postgres-based fixtures to a separate module to avoid too much clutter in conftest
